requirements.txt
torch
torchvision
Pillow
matplotlib
jupyter

Flask
